Switched most stuff to float as double precision not required.

+++ Code/C#/HVPSAPI/Messages/LiveDataMessage.cs
@@ -11,23 +11,23 @@ namespace HVPSAPI.Messages
         [JsonPropertyName(LiveDataMessageDataMemberNames.OutputVoltage)]
         [JsonInclude]
         [DataMember(Name = LiveDataMessageDataMemberNames.OutputVoltage)]
-        public double OutputVoltage { get; protected set; }
+        public float OutputVoltage { get; protected set; }
         [JsonPropertyName(LiveDataMessageDataMemberNames.OutputCurrent)]
         [JsonInclude]
         [DataMember(Name = LiveDataMessageDataMemberNames.OutputCurrent)]
-        public double OutputCurrent { get; protected set; }
+        public float OutputCurrent { get; protected set; }
         [JsonPropertyName(LiveDataMessageDataMemberNames.TotalOutputEnergy)]
         [JsonInclude]   
         [DataMember(Name = LiveDataMessageDataMemberNames.TotalOutputEnergy)]
-        public double TotalOutputEnergy { get; protected set; }
+        public float TotalOutputEnergy { get; protected set; }
         [JsonPropertyName(LiveDataMessageDataMemberNames.FirstStageVoltage)]
         [JsonInclude]
         [DataMember(Name = LiveDataMessageDataMemberNames.FirstStageVoltage)]
-        public double FirstStageVoltage { get; protected set; }
+        public float FirstStageVoltage { get; protected set; }
         [JsonPropertyName(LiveDataMessageDataMemberNames.PeakPrimaryCurrent)]
         [JsonInclude]
         [DataMember(Name = LiveDataMessageDataMemberNames.PeakPrimaryCurrent)]
-        public double PeakPrimaryCurrent { get; protected set; }
+        public float PeakPrimaryCurrent { get; protected set; }
         public LiveDataMessage()
             : base()
         {
